issue: 407928 Add support for Ubuntu14.04

+++ autogen.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
 
 #Force libnl-devel softlink to the installed libnl.so.1
-if [ ! -e /usr/lib64/libnl.so ]; then
+if [ ! -e /usr/lib64/libnl.so ] && [ -e /lib64/libnl.so.1 ]; then
 	echo Force libnl-devel softlink to the installed libnl.so.1
 	sudo ln -s /lib64/libnl.so.1  /usr/lib64/libnl.so
 fi
